==> 4_Creating_the_dataset.ipynb <==
#### Creating the dataset with matches, results and odds

package = Package('https://datahub.io/sports-data/spanish-la-liga/datapackage.json')

columns_package = ['Div','Date','HomeTeam','AwayTeam','FTHG','FTAG','FTR','HTHG','HTAG','HTR','HS','AS','HST','AST','HF','AF','HC','AC','HY','AY','HR','AR','B365H','B365D','B365A']

dictionary_variables = {'Div':'League Division','Date':'Match Date (dd/mm/yy)','Time':'Match Time','HomeTeam':'Home Team','AwayTeam':'Away Team','FTHG':'Full Time Home Team Goals','FTAG':'Full Time Away Team Goals','FTR':'Full Time Result (H Home Win, D Draw, A Away Win)','HS':'Home Team Shots','AS':'Away Team Shots','HST':'Home Team Shots on Target','AST':'Away Team Shots on Target','B365H':'Bet365 home win odds','B365D':'Bet365 draw odds','B365A':'Bet365 away win odds','season':'number of season'}

# Create historical data : season 2009-2010 to 2018-2019

historical_data = pd.DataFrame()

for resource in package.resources:
        if resource.descriptor['datahub']['type'] == 'derived/csv':
            x = pd.DataFrame(data = resource.read())
            x = x.drop(x.columns[range(25,len(x.columns))], axis=1)
            x.columns = columns_package
            x = x.drop(['HTHG','HTAG','HTR','HF','AF','HC','AC','HY','AY','HR','AR'], axis = 1)
            x['season'] = resource.name[7:11]
            historical_data = historical_data.append(x)

# Append season 2019-2020

season_1920 = pd.read_csv('https://www.football-data.co.uk/mmz4281/1920/SP1.csv', header = 0, usecols = ['Div', 'Date', 'HomeTeam', 'AwayTeam', 'FTHG', 'FTAG', 'FTR', 'HS', 'AS', 'HST', 'AST', 'B365H', 'B365D', 'B365A'])
season_1920['season'] = '1920'
historical_data = historical_data.append(season_1920)

# Current season (2020-2021)

season_2021 = pd.read_csv('http://www.football-data.co.uk/mmz4281/2021/SP1.csv', header = 0, usecols = ['Div', 'Date', 'HomeTeam', 'AwayTeam', 'FTHG', 'FTAG', 'FTR', 'HS', 'AS', 'HST', 'AST', 'B365H', 'B365D', 'B365A'])
season_2021['season'] = '2021'
historical_data = historical_data.append(season_2021)

# Changing Date type

historical_data['Date'] = pd.to_datetime(historical_data['Date'], dayfirst = True)

#### ------------------------------------------------------------------------------------------------------

#### Because the season 2020-21 is over I will predict the last 10 matches so I need to delete them from the historical_data just created.

historical_data.drop(historical_data.tail(10).index,inplace = True)

#### ------------------------------------------------------------------------------------------------------

historical_data.tail(10)

Now a clasification by season/team and date by date is needed to calculate the trends in terms of points, winning/lossing streaks, goals, shoots,etc.

# Table by season|HomeTeam|Date

aux_table_1 = pd.DataFrame(historical_data.groupby(['season','HomeTeam', 'Date'])['Div'].count()).reset_index()
aux_table_1.drop(columns =['Div'], inplace=True)
aux_table_1.rename(columns={"HomeTeam": "Team"}, inplace = True)
aux_table_1['Home/Away'] = 'H'

# Table by season|AwayTeam|Date
aux_table_2 = pd.DataFrame(historical_data.groupby(['season','AwayTeam', 'Date'])['Div'].count()).reset_index()
aux_table_2.drop(columns =['Div'], inplace=True)
aux_table_2.rename(columns={"AwayTeam": "Team"}, inplace = True)
aux_table_2['Home/Away'] = 'A'

# Table by season|Team|Date
aux_table = aux_table_1.append(aux_table_2)
aux_table.reset_index(inplace=True)
aux_table.drop(columns =['index'], inplace=True)

# Assigning number of match
aux_table.sort_values(by=['season','Team','Date'], ascending = True, inplace = True)
aux_table['match']  = aux_table.groupby(['season', 'Team']).cumcount() + 1

aux_table

Now I will add the information from the historica dataset

# Adding the variables from historical dataset

historical_acc_1 = aux_table.merge(historical_data, how = 'inner', left_on=['Date','Team'], right_on=['Date','HomeTeam'])
historical_acc_2 = aux_table.merge(historical_data, how = 'inner', left_on=['Date','Team'], right_on=['Date','AwayTeam'])
historical_acc = historical_acc_1.append(historical_acc_2)

# Rearranging

historical_acc.sort_values(by=['season_x','Team','Date'], inplace = True)

# Calculation of the points

historical_acc['victory']  = historical_acc['Home/Away'] == historical_acc['FTR'] 
historical_acc['draw'] = historical_acc['FTR'] == 'D'
historical_acc['points'] = (historical_acc['victory']) * 3 + historical_acc['draw']
                           
# Removing useless columns
                           
historical_acc.drop(labels=['HomeTeam','AwayTeam','season_y','victory','draw'], axis = 1, inplace = True)
historical_acc.rename(columns = {'season_x': 'season'}, inplace = True)

historical_acc

In the same way I calculated the points I will calculate the Goals/Shots/Shots in Target

# Calculation of the goals/shots/shots on target

historical_acc['in_home']  = historical_acc['Home/Away'] == 'H'
historical_acc['away_game']  = historical_acc['Home/Away'] == 'A'

# For

historical_acc['goals_for'] = (historical_acc['in_home'] * historical_acc['FTHG']) + (historical_acc['away_game'] * historical_acc['FTAG'])
historical_acc['shots_for'] = (historical_acc['in_home'] * historical_acc['HS']) + (historical_acc['away_game'] * historical_acc['AS'])
historical_acc['shots_target_for'] = (historical_acc['in_home'] * historical_acc['HST']) + (historical_acc['away_game'] * historical_acc['AST'])

# Against

historical_acc['goals_against'] = (historical_acc['in_home'] * historical_acc['FTAG']) + (historical_acc['away_game'] * historical_acc['FTHG'])
historical_acc['shots_against'] = (historical_acc['in_home'] * historical_acc['AS']) + (historical_acc['away_game'] * historical_acc['HS'])
historical_acc['shots_target_against'] = (historical_acc['in_home'] * historical_acc['AST']) + (historical_acc['away_game'] * historical_acc['HST'])

# Cleansing

historical_acc.reset_index(inplace = True)
historical_acc.drop(columns=['index','in_home','away_game','FTHG','FTAG','FTR','HS','AS','HST','AST'], inplace = True)
historical_acc = historical_acc[['season','Date','match','Team','Div','goals_for','goals_against','shots_for','shots_against','shots_target_for','shots_target_against','points']]

Calculating accumulated numbers

historical_acc['goals_for_acc'] = historical_acc.groupby(['season','Team'])['goals_for'].cumsum() - historical_acc['goals_for']
historical_acc['goals_against_acc'] = historical_acc.groupby(['season','Team'])['goals_against'].cumsum() - historical_acc['goals_against']
historical_acc['shots_for_acc'] = historical_acc.groupby(['season','Team'])['shots_for'].cumsum() - historical_acc['shots_for']
historical_acc['shots_against_acc'] = historical_acc.groupby(['season','Team'])['shots_against'].cumsum() - historical_acc['shots_against']
historical_acc['shots_target_for_acc'] = historical_acc.groupby(['season','Team'])['shots_target_for'].cumsum() - historical_acc['shots_target_for']
historical_acc['shots_target_against_acc'] = historical_acc.groupby(['season','Team'])['shots_target_against'].cumsum() - historical_acc['shots_target_against']
historical_acc['points_acc'] = historical_acc.groupby(['season','Team'])['points'].cumsum() - historical_acc['points']

Calculating ranking by round for each season

historical_acc['ranking'] = historical_acc.groupby(['season','match'])['points_acc'].rank(method='first', ascending=True).astype('int64')

In terms of future simplicity the column 'match' will be added to historical_data

historical_data = pd.merge(historical_data,historical_acc[['Date','Team','match']],left_on=['Date','HomeTeam'],right_on = ['Date','Team'], how='left')
historical_data = pd.merge(historical_data,historical_acc[['Date','Team','match']],left_on=['Date','AwayTeam'],right_on = ['Date','Team'], how='left',suffixes=('_home','_away'))
historical_data = historical_data[['Div','Date','season','HomeTeam','AwayTeam','match_home','match_away','B365H','B365D','B365A','FTR']]

columns_to_append= ['Date','Team','goals_for_acc','goals_against_acc','shots_for_acc','shots_against_acc','shots_target_for_acc','shots_target_against_acc','points_acc','ranking']

df_to_be_trained_home = historical_data.merge(historical_acc[columns_to_append], how = 'inner', left_on = ['Date','HomeTeam'], right_on =['Date','Team']).iloc[:,[12,13,14,15,16,17,18,19]]
df_to_be_trained_home = df_to_be_trained_home.add_suffix('_home')

df_to_be_trained_away = historical_data.merge(historical_acc[columns_to_append], how = 'inner', left_on = ['Date','AwayTeam'], right_on =['Date','Team']).iloc[:,[12,13,14,15,16,17,18,19]]
df_to_be_trained_away = df_to_be_trained_away.add_suffix('_away')


df_to_be_trained_aux = pd.concat([historical_data,df_to_be_trained_home,df_to_be_trained_away], axis = 1)
df_to_be_trained_aux.rename(columns={'FTR':'result','match_home':'match'}, inplace = True)

The final dataset to be trained will be the accumulated figures difference between the teams of that match at the time of the match

for col in df_to_be_trained_aux.columns:
    if col[-5:]  == '_home' and col[0:5] != 'match':
        df_to_be_trained_aux[col[:-5]]= df_to_be_trained_aux[col] - df_to_be_trained_aux[col[:-5]+'_away']
        df_to_be_trained_aux.drop(columns=col, axis = 1, inplace = True)
        
    elif col[-5:] =='_away' and col[0:4] != 'Team':
        
        df_to_be_trained_aux.drop(columns=col, axis = 1, inplace = True)
            
# Setting the difference in the classification df_to_be_trained_aux['ranking'] = df_to_be_trained_aux['ranking'] * - 1

columns = ['Div','season','match','HomeTeam','AwayTeam','goals_for_acc','goals_against_acc', 'shots_for_acc', 'shots_against_acc','shots_target_for_acc', 'shots_target_against_acc', 'points_acc','ranking','result','B365H','B365D','B365A']
final_dataset = df_to_be_trained_aux[columns]

final_dataset.tail()
