==> 8_Applying_algorithms.ipynb <==
#### Gathering all the prediction results

# Assigning clusters to the dataset

kmeans = KMeans(n_clusters= 3 ,**kmeans_kwargs)
kmeans.fit(X_train_scaled)
k_predictions = kmeans.predict(X_test_scaled)
k_predictions_all = kmeans.predict(features_scaled)

# Defining a dictionary with the best ML models

models = {'dtc':[predictions_dtc_all,predictions_prob_dtc_all,dtc], 'rfc':[predictions_rfc_all, predictions_prob_rfc_all,rfc], 'lr':[predictions_lr_all,predictions_prob_lr_all,lr],'lda':[predictions_lda_all,predictions_prob_lda_all,lda],'svc':[predictions_svc_all, predictions_prob_svc_all,svc]}

# Applying these models to the whole dataset

final_dataset_fin = pd.DataFrame()
    
for model in models:
    
    final_dataset_aux = historical_data.copy()
    final_dataset_aux['cluster'] = k_predictions_all
    final_dataset_aux['Prob_Home%'] = np.round(models[model][1][:,2]*100,2)
    final_dataset_aux['Prob_Draw%'] = np.round(models[model][1][:,1]*100,2)
    final_dataset_aux['Prob_Away%'] = np.round(models[model][1][:,0]*100,2)
    final_dataset_aux['Prediction'] = models[model][0]
    final_dataset_aux['Prediction'] = lab_enc_results.inverse_transform(final_dataset_aux['Prediction'])
    final_dataset_aux['hit_result'] = (final_dataset_aux['FTR'] == final_dataset_aux['Prediction'])*1
    final_dataset_aux['paid_quota'] = ((final_dataset_aux['FTR'] == 'A') * final_dataset_aux['B365A']) + ((final_dataset_aux['FTR'] == 'D') * final_dataset_aux['B365D']) + ((final_dataset_aux['FTR'] == 'H') * final_dataset_aux['B365H'])
    final_dataset_aux['Prize'] = final_dataset_aux['hit_result'] * final_dataset_aux['paid_quota']
    final_dataset_aux['Profit'] =  final_dataset_aux['Prize'] - 1
    final_dataset_aux['Difficulty'] = final_dataset_aux['paid_quota'].astype('int')
    final_dataset_aux['Model'] = model
    final_dataset_fin = final_dataset_fin.append(final_dataset_aux)

# Exporting results

final_dataset_fin.to_excel('user_report/final_dataset_fin.xlsx')

final_dataset_fin.tail()
